call it types and move it to relevant places
