+++ src/services/ckan/types/package.types.ts
