+++ src/services/ckan/types/packageSearch.types.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 PNED G.I.E.
 //
 // SPDX-License-Identifier: Apache-2.0
-import { Dataset } from './dataset.interface';
+import { Dataset } from './../../../types/dataset.types';
 
 export interface PackageSearchOptions {
   tags?: string[];
